+++ src/main/java/nl/jqno/equalsverifier/ConfiguredEqualsVerifier.java
@@ -96,6 +96,17 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
     /**
      * Factory method. For general use.
      *
+     * @param classes An iterable containing the classes for which {@code equals} method should be
+     *     tested.
+     * @return A fluent API for EqualsVerifier.
+     */
+    public MultipleTypeEqualsVerifierApi forClasses(Iterable<Class<?>> classes) {
+        return new MultipleTypeEqualsVerifierApi(ListBuilders.fromIterable(classes), this);
+    }
+
+    /**
+     * Factory method. For general use.
+     *
      * @param first A class for which the {@code equals} method should be tested.
      * @param second Another class for which the {@code equals} method should be tested.
      * @param more More classes for which the {@code equals} method should be tested.